remove list_items schema while maintaining all functionality #413
